Uninstall `wandb` from notebook environments
Due to undesired behavior in https://www.kaggle.com/code/ultralytics/yolov8/comments#2306977

+++ utils/__init__.py
@@ -61,7 +61,8 @@ def notebook_init(verbose=True):
 
     import psutil
 
-    os.system('pip uninstall -y wandb')
+    if check_requirements('wandb', install=False):
+        os.system('pip uninstall -y wandb')  # eliminate unexpected account creation prompt with infinite hang
     if is_colab():
         shutil.rmtree('/content/sample_data', ignore_errors=True)  # remove colab /sample_data directory
